Refactor custom strategy definitions into strategies.py
